% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absmarkovchain.R
\name{absmarkovchain}
\alias{absmarkovchain}
\title{Absorbing Markov Chain}
\usage{
absmarkovchain(x0, P, n)
}
\arguments{
\item{x0}{initial state, integer valued.}

\item{P}{transition probability matrix (IxJ) where the absorbing states are the last J-I rows. I = number of transitory states, J = total number of states.}

\item{n}{the number of sequences.}
}
\value{
A list where each element is a sequence of events from P.
}
\description{
This function returns n sequences of events from an absorbing Markov chain.
}
