% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amc.R
\name{amc}
\alias{amc}
\title{AMC model}
\usage{
amc(N, B = 10000, xsi)
}
\arguments{
\item{N}{A two or three dimensional array (KxIxJ) where K is the number of subgroups, I is the number of transitory states, J is the total number of states.}

\item{B}{Number of posterior draws to get (default is 10000)}

\item{xsi}{the a vector of length I where xsi is the prior value for the uniform dirichlet prior.}
}
\value{
A three dimensional array
}
\description{
This function fits an indpendent AMC model on each row of the transition matrix and outputs posterior draws of the transition probabilities.
}
