% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamc.R
\name{hamc}
\alias{hamc}
\title{HAMC model}
\usage{
hamc(
  N,
  B = 10000,
  prior.alpha = NULL,
  g.a = NULL,
  g.b = NULL,
  MCMC.cores = 1,
  burnin = 500
)
}
\arguments{
\item{N}{A two or three dimensional array of counts (KxIxJ) where K is the number of subgroups, I is the number of transitory states, J is the total number of states. Include NA's in matrix corresponding to impossible transitions.}

\item{B}{Number of posterior draws to get (default is 10000)}

\item{prior.alpha}{the a vector of length I where xsi is the prior value for the uniform dirichlet prior.}

\item{g.a}{The prior shape parameter of the gamma distribution for the gamma parameter. Default uses a mean of 20 and standard deviation of 18.}

\item{g.b}{The prior rate parameter of the gamma distribution for the gamma parameter. Default uses a mean of 20 and standard deviation of 18.}

\item{MCMC.cores}{the number of cores to use for parallel processing, the default is 1.}

\item{burnin}{The number of iterations to be removed as burnin. Default is 500.}
}
\value{
A list with the following:
\itemize{
\item alpha - Posterior draws of the alpha parameter. (BxIxJ)
\item gamma - Posterior draws of the gamma parameter. (BxI)
\item theta - Posterior draws of the transition probabilities. (BxKxIxJ)
}
}
\description{
This function fits an independent HAMC model on each row of the transition matrix and outputs posterior draws of the two hyper parameters (alpha and gamma) and the transition probabilities.
}
